project/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-ckeditor==6.4.2
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-shopping-cart==0.1
django-taggit==3.0.0
idna==3.3
Pillow==9.2.0
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.2
stripe==4.0.2
urllib3==1.26.11